Make force flag act as expected
Using the force flag used to be "download every story" before the
assumption flags were introduced, and then it acted like -y was being
used.

This commit makes it so it only forces a download after the confirmation
prompt is positive and allows the flag to be used in conjunction with
the assumption flags.